junit test case added for controller,service and repo

+++ src/main/java/com/example/demo/service/LikeServiceImpl.java
@@ -2,11 +2,16 @@ package com.example.demo.service;
 
 import com.example.demo.dto.LikeDto;
 import com.example.demo.dto.LikeRequest;
+import com.example.demo.exception.CustomFeignException;
+import com.example.demo.exception.LikeNotFoundException;
 import com.example.demo.feign.PostFeign;
 import com.example.demo.feign.UserFeign;
 import com.example.demo.model.Like;
 import com.example.demo.repo.LikeRepo;
+import com.netflix.hystrix.exception.HystrixRuntimeException;
+import feign.FeignException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
@@ -14,7 +19,8 @@ import org.springframework.stereotype.Service;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
-import static com.example.demo.constant.LikeConstant.LIKE_DELETED;
+
+import static com.example.demo.constant.LikeConstant.*;
 
 
 @Service
@@ -33,49 +39,80 @@ public class LikeServiceImpl implements LikeService {
 
 
     @Override
-    public List<LikeDto> getLikes(String postOrCommentId) {
+    public List<LikeDto> getLikes(String postOrCommentId, Integer page, Integer pageSize) {
 
-        Query query = new Query();
-        query.addCriteria(Criteria.where("postOrCommentId").is(postOrCommentId));
-        List<Like> listOfLikes = mongoTemplate.find(query, Like.class);
-        List<LikeDto> likesDto = new ArrayList<>();
-        for(Like like:listOfLikes){
-            LikeDto likeDto= new LikeDto(like.getLikeId(),like.getPostOrCommentId(),userFeign.getUserDetails(like.getLikedBy()),like.getCreatedAt());
-            likesDto.add(likeDto);
+        try {
+            if (page == null) {
+                page = 1;
+            }
+            if (pageSize == null) {
+                pageSize = 10;
+            }
+            List<Like> likes = likeRepo.findByPcId(postOrCommentId, PageRequest.of(page - 1, pageSize));
+            List<LikeDto> likeDtoList = new ArrayList<>();
+        for(Like like:likes){
+            LikeDto likeDto= new LikeDto(like.getId(),like.getPcId(),userFeign.getUserDetails(like.getLikedBy()),like.getCreatedAt());
+            likeDtoList.add(likeDto);
         }
-        return likesDto;
+            if (likeDtoList.isEmpty()) {
+                throw new LikeNotFoundException(LIKE_NOT_FOUND_FOR_POST_OR_COMMENT + postOrCommentId);
+            }
+        return likeDtoList;
+        } catch (FeignException | HystrixRuntimeException e) {
+            throw new CustomFeignException(FEIGN_EXCEPTION);
+        }
     }
 
     @Override
     public LikeDto createLike(String postOrCommentId, LikeRequest likeRequest) {
-        Like like = new Like();
-        like.setPostOrCommentId(postOrCommentId);
-        like.setLikedBy(likeRequest.getLikedBy());
-        like.setCreatedAt(new Date());
-        likeRepo.save(like);
-        return new LikeDto(like.getLikeId(),like.getPostOrCommentId(),userFeign.getUserDetails(like.getLikedBy()),like.getCreatedAt());
+        try {
+            Like like = new Like();
+            like.setPcId(postOrCommentId);
+            like.setLikedBy(likeRequest.getLikedBy());
+            like.setCreatedAt(new Date());
+            likeRepo.save(like);
+            return new LikeDto(like.getId(), like.getPcId(), userFeign.getUserDetails(like.getLikedBy()), like.getCreatedAt());
+        } catch (FeignException | HystrixRuntimeException e) {
+            throw new CustomFeignException(FEIGN_EXCEPTION);
+        }
 
     }
 
     @Override
     public LikeDto getLikeDetails(String postOrCommentId, String likeId) {
-        Like like = likeRepo.findByLikeId(likeId);
-        return new LikeDto(like.getLikeId(),like.getPostOrCommentId(),userFeign.getUserDetails(like.getLikedBy()),like.getCreatedAt());
+        try {
+            Like like = likeRepo.findByPcIdAndId(postOrCommentId, likeId);
+            if (like == null) {
+                throw new LikeNotFoundException(LIKE_NOT_FOUND_FOR_POST_OR_COMMENT + postOrCommentId + LIKE_ID + likeId);
+            }
+            return new LikeDto(like.getId(), like.getPcId(), userFeign.getUserDetails(like.getLikedBy()), like.getCreatedAt());
+        } catch (FeignException | HystrixRuntimeException e) {
+            throw new CustomFeignException(FEIGN_EXCEPTION);
+        }
 
     }
 
     @Override
     public Integer getLikesCount(String postOrCommentId) {
-        Query query = new Query();
-        query.addCriteria(Criteria.where("postOrCommentId").is(postOrCommentId));
-        List<Like> likes = mongoTemplate.find(query, Like.class);
-        return likes.size();
+        try {
+            List<Like> likes = likeRepo.findByPcId(postOrCommentId);
+            return likes.size();
+        } catch (Exception e) {
+            throw new LikeNotFoundException(LIKE_NOT_FOUND_FOR_POST_OR_COMMENT + postOrCommentId);
+        }
     }
 
     @Override
-    public String deleteLike(String likeId) {
-         likeRepo.deleteById(likeId);
-         return LIKE_DELETED;
-
+    public LikeDto removeLike(String postOrCommentId, String likeId) {
+        try {
+            Like like = likeRepo.findByPcIdAndId(postOrCommentId, likeId);
+            if (like == null) {
+                throw new LikeNotFoundException(LIKE_NOT_FOUND_FOR_POST_OR_COMMENT + postOrCommentId + LIKE_ID + likeId);
+            }
+            likeRepo.deleteById(likeId);
+            return new LikeDto(like.getId(), like.getPcId(), userFeign.getUserDetails(like.getLikedBy()), like.getCreatedAt());
+        } catch (FeignException | HystrixRuntimeException e) {
+            throw new CustomFeignException(FEIGN_EXCEPTION);
+        }
     }
 }
